// libs/features/collection/billing/src/lib/shared/enums/payment-header.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
export enum PaymentHeader {
  //Main heading: CSR, validator view.
  RECEIVE_GCC_PAYMENT = 'BILLING.RECEIVE-GCC-PAYMENT',
  RECEIVE_PAYMENT = 'BILLING.RECEIVE-PAYMENT',
  PAYMENT_NOTICE = 'BILLING.SEND-PAYMENT-NOTICE',
  RECEIVE_PAYMENT_MOF = 'BILLING.RECEIVE-PAYMENT-MOF',

  //Heading for validator
  GCC_PAYMENT_NOTICE = 'BILLING.GCC-PAYMENT-NOTICE',
  RECEIVE_PAYMENT_REJECT = 'BILLING.REJECT-RECEIVE-PAYMENT',
  RECEIVE_PAYMET_GCC_REJECT = 'BILLING.REJECT-RECEIVE-GCC-PAYMENT',
  RECEIVE_PAYMENT_MOF_REJECT = 'BILLING.REJECT-RECEIVE-PAYMENT-MOF',
  GCC_PAYMENT_NOTICE_REJECT = 'BILLING.REJECT-GCC-PAYMENT-NOTICE',
  RECEIVE_PAYMENT_RETURN = 'BILLING.RETURN-RECEIVE-PAYMENT',
  RECEIVE_PAYMET_GCC_RETURN = 'BILLING.RETURN-RECEIVE-GCC-PAYMENT',
  RECEIVE_PAYMENT_MOF_RETURN = 'BILLING.RETURN-RECEIVE-PAYMENT-MOF',
  GCC_PAYMENT_NOTICE_RETURN = 'BILLING.RETURN-GCC-PAYMENT-NOTICE',
  EST_PAYMENT_NOTICE = 'BILLING.EST-PAYMENT-NOTICE',
  EST_PAYMENT_NOTICE_RETURN = 'BILLING.RETURN-EST-PAYMENT-NOTICE',
  EST_PAYMENT_NOTICE_REJECT = 'BILLING.REJECT-EST-PAYMENT-NOTICE'
}